refactor: Restructure project to standard Next.js layout
This commit refactors the project from a monorepo structure to a standard Next.js application layout to resolve Vercel build and deployment issues.

Changes include:
- Moved the Next.js application files from `apps/web/` to the project root.
- The `app` and `components` directories are now at the top level.
- Updated all relevant import paths to use the `@/` alias for consistency and maintainability.
- Removed the now-obsolete `apps/` directory structure.
