+++ app/api/ingest/gsc/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { ingestGscData } from '../../../../../../services/ingestion/gsc-ingestor';
+import { GSCIngestionService } from '@/services/ingestion/GSCIngestionService';
 
 /**
  * API route handler to trigger the GSC data ingestion process.
@@ -19,11 +19,13 @@ export async function POST(request: NextRequest) {
             return NextResponse.json({ error: 'Dates must be in YYYY-MM-DD format.' }, { status: 400 });
         }
 
-        // Using a try-catch block for the service invocation
+        // Using a try-catch block for the service instantiation as well
         try {
+            const ingestionService = new GSCIngestionService();
+
             // Running this async, but not awaiting it, to avoid long-running serverless function timeouts.
             // For production, this should be moved to a background job queue (e.g., Google Cloud Tasks).
-            ingestGscData(siteUrl, startDate, endDate)
+            ingestionService.ingestData(siteUrl, startDate, endDate)
               .then(() => {
                 console.log(`GSC ingestion job completed successfully for ${siteUrl}.`);
               })
