+++ app/api/pages/losses/route.ts
