+++ package.json
@@ -6,32 +6,27 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "test": "jest"
+    "lint": "next lint"
   },
   "dependencies": {
-    "clsx": "^2.1.1",
-    "commander": "^12.0.0",
-    "firebase-admin": "^12.1.0",
-    "googleapis": "^137.1.0",
-    "lucide-react": "^0.540.0",
-    "next": "14.2.3",
     "react": "^18",
     "react-dom": "^18",
+    "next": "14.2.3",
+    "firebase-admin": "^12.1.0",
+    "googleapis": "^137.1.0",
+    "commander": "^12.0.0",
     "recharts": "^2.12.7",
-    "tailwind-merge": "^2.3.0"
+    "tailwind-merge": "^2.3.0",
+    "clsx": "^2.1.1"
   },
   "devDependencies": {
-    "@types/jest": "^30.0.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.3",
-    "jest": "^30.0.5",
+    "typescript": "^5",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "ts-jest": "^29.4.1",
-    "typescript": "^5"
+    "eslint": "^8",
+    "eslint-config-next": "14.2.3"
   }
 }
