+++ services/ingestion/README.md
@@ -4,33 +4,13 @@
 
 This module is responsible for connecting to the Google Search Console (GSC) API, fetching daily performance data, normalizing it, and persisting it to a Firestore collection named `gsc_raw`.
 
-The core logic has been refactored into a standalone function, `ingestGscData`, located in `gsc-ingestor.ts`. This module is designed to be robust, with built-in retries for both API calls and database writes.
-
 It provides two ways to trigger the ingestion process: a command-line interface (CLI) and a protected HTTP endpoint.
 
 ## Setup
 
-Before running the service, you must set two environment variables containing Base64-encoded strings of your Google Cloud Service Account JSON keys.
+Before running the service, you must set the following environment variable:
 
--   `GSC_SERVICE_ACCOUNT_BASE64`: Service account credentials with **read-only access to the Google Search Console API**.
--   `FIREBASE_ADMIN_SDK_JSON_BASE64`: Service account credentials with **write access to your Firestore database**.
-
-*Note: You can use the same service account for both if it has the required permissions for both services.*
-
-## Firestore Data Schema
-
-Data is written to the `gsc_raw` collection with the following schema:
-
--   `site` (string): The GSC property URL (e.g., "sc-domain:example.com").
--   `url` (string): The normalized URL of the page.
--   `query` (string): The search query.
--   `date` (string): The date of the data in `YYYY-MM-DD` format.
--   `impressions` (number): The number of impressions.
--   `clicks` (number): The number of clicks.
--   `position` (number): The average ranking position.
--   `device` (string): The device type (e.g., "DESKTOP", "MOBILE").
--   `country` (string): The three-letter country code (e.g., "USA").
--   `searchAppearance` (string): The search appearance type (e.g., "News", "Video").
+-   `FIREBASE_ADMIN_SDK_JSON_BASE64`: A Base64-encoded string of your Google Cloud Service Account JSON key. This service account needs access to both the Google Search Console API and your Firebase/Firestore project.
 
 ## Usage
